# [Infra][add/remove topo] Improve `vm_topology` performance