@mssonicbld 2024-12-27
@lolyu PR conflicts with 202405 branch