@mssonicbld 2024-12-27
Cherry-pick PR to 202411: https://github.com/sonic-net/sonic-mgmt/pull/16245