@davidm-arista 2024-12-27
This is mixing threading with the underlying ansible multi-processing, and is a recipe for deadlocks.  Please note from https://docs.ansible.com/ansible/latest/dev_guide/developing_api.html that Ansible is not thread safe.