@lolyu 2025-01-02
> This is mixing threading with the underlying ansible multi-processing, and is a recipe for deadlocks. Please note from https://docs.ansible.com/ansible/latest/dev_guide/developing_api.html that Ansible is not thread safe.

Hi @davidm-arista, this doc: https://docs.ansible.com/ansible/latest/dev_guide/developing_api.html is about the Ansible Python API. AFAIK, sonic-mgmt doesn't have any usecase of this Python API to call `vm_topology`. Do you have any usecase that have the deadlock situation?